chore: removing unused code
