+++ client-sdks/client-js/src/client.ts
@@ -47,8 +47,6 @@ import type {
   GetMemoryConfigParams,
   GetMemoryConfigResponse,
   GetMemoryThreadMessagesResponse,
-  GetScoreByIdParams,
-  ClientScoreRowData,
 } from './types';
 import { base64RuntimeContext, parseClientRuntimeContext } from './utils';
 
@@ -629,16 +627,6 @@ export class MastraClient extends BaseResource {
     return this.request(
       `/api/scores/entity/${encodeURIComponent(entityType)}/${encodeURIComponent(entityId)}${queryString ? `?${queryString}` : ''}`,
     );
-  }
-
-  /**
-   * Retrieves a single score by ID
-   * @param params - Parameters containing the score ID
-   * @returns Promise containing the score data
-   */
-  public getScoreById(params: GetScoreByIdParams): Promise<ClientScoreRowData> {
-    const { id } = params;
-    return this.request(`/api/scores/${id}`);
   }
 
   /**
