+++ client-sdks/client-js/src/types.ts
@@ -525,10 +525,6 @@ export interface GetScoresBySpanParams {
   page?: number;
   perPage?: number;
 }
-export interface GetScoreByIdParams {
-  id: string;
-}
-
 export interface SaveScoreParams {
   score: Omit<ScoreRowData, 'id' | 'createdAt' | 'updatedAt'>;
 }
