+++ packages/deployer/src/server/handlers/routes/scores/router.ts
@@ -8,7 +8,6 @@ import {
   getScoresByRunIdHandler,
   getScoresByScorerIdHandler,
   getScoresByEntityIdHandler,
-  getScoreByIdHandler,
   saveScoreHandler,
 } from './handlers';
 
@@ -158,32 +157,6 @@ export function scoresRouter(bodyLimitOptions: BodyLimitOptions) {
       },
     }),
     getScoresByEntityIdHandler,
-  );
-
-  router.get(
-    '/:id',
-    describeRoute({
-      description: 'Get a single score by ID',
-      tags: ['scores'],
-      parameters: [
-        {
-          name: 'id',
-          in: 'path',
-          required: true,
-          schema: { type: 'string' },
-          description: 'ID of the score',
-        },
-      ],
-      responses: {
-        200: {
-          description: 'Score data',
-        },
-        404: {
-          description: 'Score not found',
-        },
-      },
-    }),
-    getScoreByIdHandler,
   );
 
   router.post(
