+++ packages/playground-ui/src/domains/scores/hooks/use-scorers.tsx
@@ -86,31 +86,3 @@ export const useScorers = () => {
     gcTime: 0,
   });
 };
-
-export const useScoreById = (id: string) => {
-  const client = useMastraClient();
-  const [score, setScore] = useState<ClientScoreRowData | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchScore = async () => {
-      setIsLoading(true);
-      try {
-        const res = await client.getScoreById({ id });
-        setScore(res);
-        setIsLoading(false);
-      } catch (error) {
-        setScore(null);
-        console.error('Error fetching score', error);
-        toast.error('Error fetching score');
-        setIsLoading(false);
-      }
-    };
-
-    if (id) {
-      fetchScore();
-    }
-  }, [id]);
-
-  return { score, isLoading };
-};
